+++ Sources/StreamChatSwiftUI/ColorPalette.swift
@@ -8,7 +8,8 @@ import UIKit
 /// Provides the colors used throughout the SDK.
 public struct ColorPalette {
     public init() {
-        // Public init.
+        navigationTitle = text
+        navigationSubtitle = textLowEmphasis
     }
 
     public var tintColor: Color = .accentColor
@@ -87,6 +88,18 @@ public struct ColorPalette {
     public lazy var composerInputBackground: UIColor = background
     public lazy var composerInputHighlightedBorder: UIColor = innerBorder
 
+    // MARK: - Navigation Bar
+    
+    public var navigationTitle: UIColor {
+        didSet {
+            let attributes: [NSAttributedString.Key: Any] = [.foregroundColor: navigationTitle]
+            UINavigationBar.appearance().titleTextAttributes = attributes
+            UINavigationBar.appearance().largeTitleTextAttributes = attributes
+        }
+    }
+
+    public var navigationSubtitle: UIColor
+    
     // MARK: - Threads
 
     public var bannerBackgroundColor: UIColor = .streamDarkGray